pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tigerfeat"
version = "0.1.0"
description = "Lightweight feature extraction wrapper for timm vision models"
readme = "README.md"
authors = [{name = "TigerFeat"}]
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch",
    "timm",
    "Pillow",
]

[project.urls]
Homepage = "https://github.com/example/tigerfeat"
